Update sensor.py
On my energymanager I had to remove the lines with "BatteryConverterDevice.DEVICE_CLASS," to get this running. During Init it complained the function was provided with 7 arguments but only takes 6. I figured it must have been these lines which are wrong. Now it runs fine. Nice work on the integration!